fix(lint): split routes/http.rs under the 500-line gate
